Don't recommend users follow themselves (readme)
Admittedly this is an extremely rare case, but previously if a user wasn't
following anyone and happened to be in the top 3 most followed users on the site
(that aren't blocked by or blocking the user) they would be recommended to
themselves in the "Who to follow" panel.

Really this only applies to @crabber right now but still. Edge cases and that.